Update on "[WIP] in memory checkpoint example"
The problem:

Reading from disk / network file system is expensive. If we can read from memory then loading checkpoints is a lot faster and we can do it more frequently.

The idea:

1. Keep a process alive which has the checkpoint in memory 
2. On cases where the Trainer dies (the host does not die), restart training, reconfigure the process group, retrieve the checkpoint from memory.
3. Reshard / replicate as necessary using DCP. Use TorchFTs pg transport which uses P2P ops.

Look at https://github.com/pytorch/pytorch/pull/149202/files#diff-b596c9c510c9a06e13d1c200dfb294b3dfd7a53eb02c3e718ae36aa718fa1326R358 for a rough example.

----------

Brainstorming docs: https://fburl.com/gdoc/w6x32v9a, https://fburl.com/gdoc/se7kh86g
Potential impact and savings: https://fburl.com/gdoc/5pcd2lkm


cc awgu kwen2501 wanchaol fegin fduwjj wz337 wconstab d4l3k c-p-i-o

[ghstack-poisoned]
