+++ torch/distributed/checkpoint/_in_memory_checkpoint.py
@@ -1,20 +1,14 @@
 import logging
-import mmap
-import os
 import pickle
-import tempfile
-import socket
-from multiprocessing import shared_memory
 import time
+from collections.abc import Generator
 from contextlib import contextmanager
 from dataclasses import dataclass
 from datetime import timedelta
-from typing import Callable, cast, Generator, Optional, TypeVar, Union
+from multiprocessing import shared_memory
+from typing import Callable, cast, Optional, TypeVar, Union
 
-import time
 import torch
-import torch.distributed as dist
-import torch.nn as nn
 from torch.distributed import ProcessGroup, Work
 from torch.distributed.tensor import _DTensorSpec, DTensor
 from torch.utils._pytree import (
@@ -23,6 +17,7 @@ from torch.utils._pytree import (
     tree_unflatten,
     TreeSpec,
 )
+
 
 logger: logging.Logger = logging.getLogger(__name__)
 
@@ -156,9 +151,9 @@ def _cast_tensor(tensor: torch.Tensor, dtype: torch.dtype) -> torch.Tensor:
     caveat that the cast tensor may be larger than the original tensor due to
     the differences in striding.
     """
-    assert (
-        type(tensor) is torch.Tensor
-    ), f"can only cast standard tensors not {type(tensor)}"
+    assert type(tensor) is torch.Tensor, (
+        f"can only cast standard tensors not {type(tensor)}"
+    )
     storage = tensor.untyped_storage()
     ret = torch.tensor(storage, dtype=dtype, device=tensor.device)
     assert ret.untyped_storage() is storage, "storage should be the same"
@@ -167,10 +162,11 @@ def _cast_tensor(tensor: torch.Tensor, dtype: torch.dtype) -> torch.Tensor:
 
 class InMemoryStorage:
     """
-    In memory storage for checkpointing. This is a simple in memory storage which 
-    uses shared memory to store the checkpoint. 1 process keeps this alive and other 
+    In memory storage for checkpointing. This is a simple in memory storage which
+    uses shared memory to store the checkpoint. 1 process keeps this alive and other
     processes can access it. TODO: can also save the memory to temporary files and load it via mmap
     """
+
     def __init__(self):
         self._storage = {}
 
@@ -304,9 +300,9 @@ class PGTransport:
                 if isinstance(inplace, DTensor):
                     inplace = inplace._local_tensor
                 t = _cast_tensor(inplace, torch.uint8)
-                assert (
-                    t.nbytes == v.nbytes
-                ), "inplace tensor storage must be the same size"
+                assert t.nbytes == v.nbytes, (
+                    "inplace tensor storage must be the same size"
+                )
             else:
                 t = torch.empty(v.nbytes, dtype=torch.uint8, device=self._device)
 
@@ -342,73 +338,3 @@ class PGTransport:
             work.wait(timeout)
 
         return tree_unflatten(values, meta.treespec)
-
-
-class ToyModel(nn.Module):
-    def __init__(self):
-        super(ToyModel, self).__init__()
-        self.net1 = nn.Linear(16, 16)
-        self.relu = nn.ReLU()
-        self.net2 = nn.Linear(16, 8)
-
-    def forward(self, x):
-        return self.net2(self.relu(self.net1(x)))
-
-if __name__ == "__main__":
-    # Example usage: torchrun --nproc-per-node=3 torch/distributed/checkpoint/_in_memory_checkpoint.py
-    rank = int(os.environ.get("RANK", "0"))
-    world_size = int(os.environ.get("WORLD_SIZE", "1"))
-    dist.init_process_group("gloo", rank=rank, world_size=world_size)
-    pg = dist.new_group(backend="gloo")
-    transport = PGTransport(
-        pg, timeout=timedelta(seconds=30), device=torch.device("cpu")
-    )
-    if rank == 0:
-        # keep this process alive, the storage in memory
-        model = ToyModel()
-        storage = InMemoryStorage()
-        storage.save("sd", model.state_dict())
-        print(f"Rank 0: Saved state_dict. {model.state_dict().keys()=}")
-
-        # Set up a server socket
-        server_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        server_socket.bind(('localhost', 12345))
-        server_socket.listen(1)
-        print("Rank 0: Server listening for connections...")
-        conn, addr = server_socket.accept()
-        print(f"Rank 0: Connection from {addr}")
-        # Send shared memory name and size
-        shm_name, data_size = storage._storage["sd"]
-        conn.sendall(f"{shm_name},{data_size}".encode('utf-8'))
-        conn.close()
-        storage.clear()
-
-    if rank == 1:
-        # TODO: waiting for server to start up
-        time.sleep(3)
-        # Connect to the server
-        client_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        client_socket.connect(('localhost', 12345))
-        # Receive shared memory name and size
-        data = client_socket.recv(1024).decode('utf-8')
-        shm_name, data_size = data.split(',')
-        data_size = int(data_size)
-        client_socket.close()
-
-        # Access the shared memory
-        storage = InMemoryStorage()
-        storage._storage["sd"] = (shm_name, data_size)
-        state_dict = storage.load("sd")
-
-        print(f"Rank 1: Checkpoint loaded. {state_dict.keys()=}")
-        transport.send_checkpoint(
-            dst_ranks=[2], step=1, state_dict=state_dict, timeout=timedelta(seconds=30)
-        )
-        print("Rank 1: Checkpoint sent and saved.")
-    elif rank == 2:
-        # Rank 1 receives the checkpoint
-        received_state_dict = transport.recv_checkpoint(
-            src_rank=1, metadata="", step=1, timeout=timedelta(seconds=30)
-        )
-        print("Rank 2: Checkpoint received")
-    dist.destroy_process_group()
